refactor(#606): move 2 controllers
